update android sdk

+++ android_sdk/include/applib/ovxinc/include/utils/vsi_nn_dtype_util_prv.h
@@ -194,11 +194,11 @@ static inline int32_t fp32_to_dfp
     type_get_range( type, &max_range, &min_range );
     if( fl > 0 )
     {
-        data = (int32_t)vsi_rint( in * (float)( 1 << fl ) );
+        data = (int32_t)vsi_rint( in * (float)( (int64_t)1 << fl ) );
     }
     else
     {
-        data = (int32_t)vsi_rint( in * ( 1.0f / (float)( 1 << -fl ) ) );
+        data = (int32_t)vsi_rint( in * ( 1.0f / (float)( (int64_t)1 << -fl ) ) );
     }
     data = vsi_nn_min( data, (int32_t)max_range );
     data = vsi_nn_max( data, (int32_t)min_range );
@@ -215,11 +215,11 @@ static inline float dfp_to_fp32
     float result;
     if( fl > 0 )
     {
-        result = (float)val * ( 1.0f / ( (float) ( 1 << fl ) ) );
+        result = (float)val * ( 1.0f / ( (float) ( (int64_t)1 << fl ) ) );
     }
     else
     {
-        result = (float)val * ( (float) ( 1 << -fl ) );
+        result = (float)val * ( (float) ( (int64_t)1 << -fl ) );
     }
     return result;
 } /* dfp_to_fp32() */
